Tidy up code and remove unnecessary returns
Blackbox the permissions logic into `sentByAdmin(msg)`, in case this logic changes later.

Remove unnecessary return statements in functions that don't return anything by tidying up the if...else if...else logic. This is fine since when a command function is called in this project, its return value is discarded, whatever it may be.

Add codeblocks to if...else if...else statements to make code more readable and less prone to bugs in the future.

Replaced an instance of `var` with `let`.